// src/store/actions/education/education-types.ts
export enum EducationActionTypes {
  GetAllRootNodes = '[Education] GetAllRootNodes',
  GetAllRootNodesSuccess = '[Education] GetAllRootNodes success',
  GetAllRootNodesFailure = '[Education] GetAllRootNodes failure',

  GetSubNodes = '[Education] GetSubNodes',
  GetSubNodesSuccess = '[Education] GetSubNodes success',
  GetSubNodesFailure = '[Education] GetSubNodes failure',

  GetLeafs = '[Education] GetLeafs',
  GetLeafsSuccess = '[Education] GetLeafs success',
  GetLeafsFailure = '[Education] GetLeafs failure',
}